flytt interfaces, m.m. til egen utils fil
